requirements.txt
# ── Core ------------------------------------------------------------------
faiss-cpu==1.11.0.post1
torch==2.3.0
sentence-transformers==5.0.0
transformers==4.54.1
huggingface-hub==0.34.3
tokenizers==0.21.4

# ── Data handling ---------------------------------------------------------
numpy==1.26.4
scipy==1.16.1
pandas==2.3.1
scikit-learn==1.7.1
joblib==1.5.1
threadpoolctl==3.6.0

# ── Streamlit UI & API ----------------------------------------------------
streamlit==1.47.1
fastapi==0.110.0
uvicorn[standard]==0.22.0
altair==5.5.0
blinker==1.9.0
watchdog==6.0.0
pydeck==0.9.1

# ── I/O & utilities -------------------------------------------------------
attrs==25.3.0
click==8.2.1
filelock==3.18.0
requests==2.32.4
tqdm==4.67.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rich==13.7.1
python-dotenv==0.20.0
slowapi==0.1.9

# ── PDF & doc extraction --------------------------------------------------
PyMuPDF==1.24.2
python-docx==1.1.2
html2text==2020.1.16
pdfminer.six==20221105
Markdown==3.4.1
ebooklib==0.18
beautifulsoup4==4.12.3

# ── Testing ---------------------------------------------------------------
pytest==8.0.0
black==23.3.0
isort==5.12.0

# ── LLM & Summarization ---------------------------------------------------
langchain==0.1.4
langchain-community==0.0.16

# ── Documentation ---------------------------------------------------------
mkdocs==1.5.2
mkdocs-material==9.1.21